+++ src/Framework/Gui/Bar.hpp
@@ -26,7 +26,7 @@ public:
 
 	sf::IntRect const &GetRect() { return m_frame.GetRect(); }
 
-	void SetSmoothTransformationTime(float _time) { baseAnimationTime = _time; currentAnimationTime = _time; }
+	void SetSmoothTransformationTime(double _time) { baseAnimationTime = _time; currentAnimationTime = _time; }
 
 	//sets the font
 	void SetText(sf::Font const &_font, sf::Color const &_color, int _size);
@@ -34,7 +34,7 @@ public:
 	//sets the character size
 	void SetCharacterSize(int _size) { m_text.setCharacterSize(_size); }
 
-	void Update(float _elapsedTime);
+	void Update(double _elapsedTime);
 
 	//Renders the beam
 	void Render(sf::RenderTarget &_target, bool _withNumbers = false);
@@ -56,12 +56,12 @@ private:
 	int lastValue;
 
 	int smoothAnimationBaseValue;
-	float baseAnimationTime;
-	float currentAnimationTime;
+	double baseAnimationTime;
+	double currentAnimationTime;
 
 	sf::Text m_text;                    //the text
 
 	void StartAnimation();
-	void HandleAnimation(float _elapsedTime);
+	void HandleAnimation(double _elapsedTime);
 
 };
